feat: Write only columns containing sample names in "Sample group"
Columns retrieved for "Sample group" blocks now must contain the group
tag and a sample name. Columns not containing a sample name are ignored.
For example, using the tag "Intensity" will no longer include columns
such as "Intensity" or "Intensity total". In order to write these
columns to the report, they must be manually specified in a
"Feature group" block.